Cut back some code
Reduced the amount of code significantly in RenderObjectDataFormat by removing use of enums. This significantly reduces the amount of overall code and work in the rendering components

+++ crispinmobile/src/main/java/com/games/crispin/crispinmobile/Utilities/OBJModelLoader.java
@@ -473,9 +473,9 @@ public class OBJModelLoader
                 break;
         }
 
-        renderObjectData.setNumPositionComponents(numberPositionDataElements);
-        renderObjectData.setNumNormalComponents(numberNormalDataElements);
-        renderObjectData.setNumTexelComponents(numberTexelDataElements);
+        renderObjectData.setNumPositionComponents((byte)numberPositionDataElements);
+        renderObjectData.setNumNormalComponents((byte)numberNormalDataElements);
+        renderObjectData.setNumTexelComponents((byte)numberTexelDataElements);
         return renderObjectData.processData();
     }
 }
